avoid some left shifting to sign bit
doesn't patch out left shifting negative values as that seems to be
properly handled by gcc/clang/msvc (even while the standard makes it
undefined), and patching it would uglify the code too much

+++ plugins/gpu_neon/psx_gpu/psx_gpu_simd.c
@@ -1851,7 +1851,7 @@ void setup_spans_up_down(psx_gpu_struct *psx_gpu, vertex_struct *v_a,
   vec_8x8s dither_offsets_short;                                               \
                                                                                \
   dither_row =                                                                 \
-   (dither_row >> dither_shift) | (dither_row << (32 - dither_shift));         \
+   (dither_row >> dither_shift) | ((u64)dither_row << (32 - dither_shift));    \
   gvdup_n_u32(dither_offsets_short, dither_row);                               \
   setup_blocks_span_initialize_dithered_##texturing()                          \
